Feat #406 : Getting local models is now centralised LocalChatModelFac…

+++ core/src/main/java/com/devoxx/genie/model/jan/Data.java
@@ -15,7 +15,7 @@ public class Data {
     private String object;
 
     @JsonProperty("ctx_len")
-    private Integer ctxLen;
+    private Long ctxLen;
 
     @JsonProperty("max_tokens")
     private Integer maxTokens;
